✨ Change y_pos from float to double in gameplayer struct for improved precision

+++ src/server/include/GamePlayer.hpp
@@ -5,7 +5,7 @@
 typedef struct gameplayer {
     int id;
     char username[20];
-    float y_pos;
+    double y_pos;
     double velocity;
     int coins_collected;
     bool is_dead;
